requirements.txt
django-redis==5.4.0
argon2-cffi==23.1.0
Django==5.1.2
Pillow==11.0.0
Djangorestframework==3.15.2
psycopg2==2.9.10 
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-redis==5.4.0
redis==5.2.0 
Hiredis==3.0.0
#django-cacheops==7.1 
whitenoise==6.8.2
django-environ==0.11.2
dj-database-url==2.3.0
django-csp==3.8
# channels==4.2.0
gunicorn==23.0.0
cloudinary==1.42.1
django-cloudinary-storage==0.3.0